Move proptypes missed in last commit
