+++ src/app/components/Headline/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { extractText } from '../../helpers/blockHandlers';
-import { textPropTypes, textDefaultPropTypes } from '../../helpers/proptypes';
+import { textPropTypes, textDefaultPropTypes } from '../../models/proptypes';
 import { C_EBON, FF_NEWS_SANS_REG } from '../../../lib/constants/styles';
 
 const StyledHeadline = styled.h1`
