+++ src/app/components/SubHeading/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { extractText } from '../../helpers/blockHandlers';
-import { textPropTypes, textDefaultPropTypes } from '../../helpers/proptypes';
+import { textPropTypes, textDefaultPropTypes } from '../../models/proptypes';
 
 const SubHeading = ({ blocks }) => {
   const { text } = extractText(blocks);
